Refactor `onBackPressed` to use `OnBackPressedDispatcher`
This commit refactors the `onBackPressed` method in `EntriesActivity` to use `OnBackPressedDispatcher.addCallback`.

It also changes the behavior of the back button to pop the fragment back stack if there is more than one entry, instead of finishing the activity.

Additionally, it initializes `selectedText` in `BookmarkDetailViewModel` to an empty string instead of null, and fixes a typo in a URL.